+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## [1.9.0](https://github.com/tensorplex-labs/dojo/compare/v1.8.0...v1.9.0) (2025-03-20)
+
+### Features
+
+* analytics API ([#114](https://github.com/tensorplex-labs/dojo/issues/114)) ([#162](https://github.com/tensorplex-labs/dojo/issues/162)) ([619937c](https://github.com/tensorplex-labs/dojo/commit/619937cb12b6860d609824232f0ff14e610f48b6))
+
+## [1.8.0](https://github.com/tensorplex-labs/dojo/compare/v1.7.2...v1.8.0) (2025-03-12)
+
+### Features
+
+* allow dual mining ([#151](https://github.com/tensorplex-labs/dojo/issues/151)) ([6952e54](https://github.com/tensorplex-labs/dojo/commit/6952e543231fda12efdb3a3b946ad80806e610ab))
+* shutdown validator when syn-API is down ([#132](https://github.com/tensorplex-labs/dojo/issues/132)) ([e8d5a33](https://github.com/tensorplex-labs/dojo/commit/e8d5a330c69cd4676e3ee0df6a510c3ad2c4f8d7))
+
+### Bug Fixes
+
+* fix async issues and update score column correctly ([#143](https://github.com/tensorplex-labs/dojo/issues/143)) ([2c39e0a](https://github.com/tensorplex-labs/dojo/commit/2c39e0a370a7b72e123b81778c27b31940f07241)), closes [#153](https://github.com/tensorplex-labs/dojo/issues/153)
+
+## [1.7.2](https://github.com/tensorplex-labs/dojo/compare/v1.7.1...v1.7.2) (2025-03-06)
+
+### Bug Fixes
+
+* hotfix for miner's logs ([#149](https://github.com/tensorplex-labs/dojo/issues/149)) ([9bd225a](https://github.com/tensorplex-labs/dojo/commit/9bd225ab0fef21e8d6c25255ca5253dd25d086e3))
+
+## [1.7.1](https://github.com/tensorplex-labs/dojo/compare/v1.7.0...v1.7.1) (2025-03-03)
+
+### Bug Fixes
+
+* removed dojo-task-id check in blacklist func ([#147](https://github.com/tensorplex-labs/dojo/issues/147)) ([9b1a28e](https://github.com/tensorplex-labs/dojo/commit/9b1a28e4edcb603978564178f1a88109b0bc64d7))
+
+## [1.7.0](https://github.com/tensorplex-labs/dojo/compare/v1.6.6...v1.7.0) (2025-02-25)
+
+### Features
+
+* docker revamp ([#119](https://github.com/tensorplex-labs/dojo/issues/119)) ([6e36f82](https://github.com/tensorplex-labs/dojo/commit/6e36f82c9a819c6fdf4cd4de3c0eaae06ac49270))
+
+### Bug Fixes
+
+* fix logging endpoint ([#133](https://github.com/tensorplex-labs/dojo/issues/133)) ([91b94fa](https://github.com/tensorplex-labs/dojo/commit/91b94fa14d5c4286d1ae808066ab06b5cf35d5f4))
+* use effective stake (root + alpha), set 5000 for effective stake ([#142](https://github.com/tensorplex-labs/dojo/issues/142)) ([9d151fc](https://github.com/tensorplex-labs/dojo/commit/9d151fcab607fb03c3b8faeb6ca7f9d5de7228a9))
+
 ## [1.6.6](https://github.com/tensorplex-labs/dojo/compare/v1.6.5...v1.6.6) (2025-02-15)
 
 ### Bug Fixes